Ellie-GPT/pyproject.toml
[project]
name = "ellie-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
]
